requirements.txt
# Base
psycopg2-binary==2.9.9
PyYAML==6.0.2
schedule==1.2.1
requests==2.32.3

# Outils de log / utilitaires
rich==13.7.0
tabulate==0.9.0

# Tests
pytest==8.3.2
pytest-cov==5.0.0

# DevOps / Automatisation (optionnel)
ansible==10.1.0
paramiko==3.5.0

# Monitoring / Reporting
jinja2==3.1.4
markdown2==2.5.0

# Sécurité et audits
cryptography==43.0.1

# Lint / Qualité
black==24.8.0
flake8==7.1.1
